Get latest RStudio version number

Get latest RStudio version number

latest_version(
  type = c("desktop", "server"),
  stable = TRUE,
  pro = FALSE,
  os = NULL,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

type
Either "desktop" for RStudio Desktop or "server" for RStudio Server release metadata.

stable
Set to FALSE to retrieve release metadata of RStudio preview builds instead of stable builds.

pro
TRUE for the proprietary RStudio (Server) Pro edition and FALSE for the open-source RStudio (Server) edition.

os
The OS codename for which the RStudio version was built. If NULL, it will be auto-detected for the current system.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A numeric version.

Examples

latest_version(os = "macos",
               use_cache = FALSE)
#> [1] ‘2022.2.2.485’