Get RStudio release metadata

Get RStudio release metadata

releases(
  type = c("desktop", "server"),
  stable = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

type
Either "desktop" for RStudio Desktop or "server" for RStudio Server release metadata.

stable
Set to FALSE to retrieve release metadata of RStudio preview builds instead of stable builds.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

releases(type = "server",
         cache_lifespan = "1 year 2 months 3 weeks 4 days 5 hours 6 minutes 7 seconds")
#> # A tibble: 999 × 8
#>    Key                LastModified ETag  Size  StorageClass is_pro os    version
#>    <chr>              <chr>        <chr> <chr> <chr>        <lgl>  <chr> <chr>  
#>  1 rstudio-server-0.… 2011-01-19T… "\"2… 5510… STANDARD     FALSE  NA    NA     
#>  2 rstudio-server-0.… 2011-01-19T… "\"9… 5437… STANDARD     FALSE  NA    NA     
#>  3 rstudio-server-0.… 2011-01-19T… "\"3… 5620… STANDARD     FALSE  NA    NA     
#>  4 rstudio-server-0.… 2011-01-19T… "\"8… 5732… STANDARD     FALSE  NA    NA     
#>  5 rstudio-server-0.… 2011-01-20T… "\"6… 5513… STANDARD     FALSE  NA    NA     
#>  6 rstudio-server-0.… 2011-01-20T… "\"2… 5446… STANDARD     FALSE  NA    NA     
#>  7 rstudio-server-0.… 2011-01-20T… "\"2… 5611… STANDARD     FALSE  NA    NA     
#>  8 rstudio-server-0.… 2011-01-20T… "\"7… 5736… STANDARD     FALSE  NA    NA     
#>  9 rstudio-server-0.… 2011-01-20T… "\"2… 5519… STANDARD     FALSE  NA    NA     
#> 10 rstudio-server-0.… 2011-01-20T… "\"a… 5451… STANDARD     FALSE  NA    NA     
#> # … with 989 more rows