Get RStudio release metadata

Get RStudio release metadata

Usage

releases(
  type = c("desktop", "server"),
  stable = TRUE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

type

Either "desktop" for RStudio Desktop or "server" for RStudio Server release metadata.

stable

Set to FALSE to retrieve release metadata of RStudio preview builds instead of stable builds.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

releases(type = "server",
         cache_lifespan = "1 year 2 months 3 weeks 4 days 5 hours 6 minutes 7 seconds")
#> # A tibble: 999 × 8
#>    key                 last_modified       e_tag                                  size     storage_class is_pro os    version
#>    <chr>               <dttm>              <chr>                                  <chr>    <chr>         <lgl>  <chr> <chr>  
#>  1 RStudio-0.91.10.dmg 2011-01-19 13:38:35 "\"dd5e51e3b6b85a457eeec338789152af\"" 39947962 STANDARD      FALSE  NA    0.91.10
#>  2 RStudio-0.91.10.exe 2011-01-19 13:38:32 "\"fd4f8937a1f5dac517bd12cfce360aa3\"" 15760884 STANDARD      FALSE  NA    0.91.10
#>  3 RStudio-0.91.12.dmg 2011-01-20 18:18:59 "\"700587478a61c259e06451c5b2cb1df3\"" 39958093 STANDARD      FALSE  NA    0.91.12
#>  4 RStudio-0.91.12.exe 2011-01-20 18:18:56 "\"b28e69b0a0585af856bb28d0f4721013\"" 15760314 STANDARD      FALSE  NA    0.91.12
#>  5 RStudio-0.91.13.dmg 2011-01-20 20:06:01 "\"7b18caab610da14a6639f4a1e21cecdd\"" 39957534 STANDARD      FALSE  NA    0.91.13
#>  6 RStudio-0.91.13.exe 2011-01-20 20:05:57 "\"97d84682a5f565e345d537f629520bed\"" 15763302 STANDARD      FALSE  NA    0.91.13
#>  7 RStudio-0.91.21.dmg 2011-01-28 12:47:15 "\"245f1b8a8e1499e4fb72eba4b9a7b1c1\"" 39960067 STANDARD      FALSE  NA    0.91.21
#>  8 RStudio-0.91.21.exe 2011-01-28 12:47:11 "\"4093c8f165e7306f9fcafef6cbc597cc\"" 15765847 STANDARD      FALSE  NA    0.91.21
#>  9 RStudio-0.91.31.dmg 2011-01-31 10:58:32 "\"5a085876971f752a0dc2f766389e9c96\"" 39961535 STANDARD      FALSE  NA    0.91.31
#> 10 RStudio-0.91.31.exe 2011-01-31 10:58:29 "\"679c9116ac021ed9093b8b3f8f90bb6e\"" 15764106 STANDARD      FALSE  NA    0.91.31
#> # … with 989 more rows