Get RStudio release metadata

Get RStudio release metadata

Usage

releases(
  type = c("desktop", "server"),
  stable = TRUE,
  use_cache = TRUE,
  max_cache_age = "1 day"
)

Arguments

type

Either "desktop" for RStudio Desktop or "server" for RStudio Server release metadata.

stable

Set to FALSE to retrieve release metadata of RStudio preview builds instead of stable builds.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of max_cache_age.

max_cache_age

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Use Inf to disable cache expiry. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

releases(type = "server",
         max_cache_age = "1 year 2 months 3 weeks 4 days 5 hours 6 minutes 7 seconds")
#> # A tibble: 999 × 8
#>    key                               last_modified       e_tag                                  size    storage_class is_pro os    version
#>    <chr>                             <dttm>              <chr>                                  <chr>   <chr>         <lgl>  <chr> <chr>  
#>  1 rstudio-server-0.91.10-amd64.deb  2011-01-19 13:38:41 "\"23eff550ee2d2a0a4663806ba96d3a56\"" 5510286 STANDARD      FALSE  NA    0.91.10
#>  2 rstudio-server-0.91.10-i386.deb   2011-01-19 13:38:42 "\"9a43cdc2d9309fb72465cb3eac9bf052\"" 5437970 STANDARD      FALSE  NA    0.91.10
#>  3 rstudio-server-0.91.10-i686.rpm   2011-01-19 13:38:44 "\"3c4e547285c6f563e9d9bf3510d407b5\"" 5620064 STANDARD      FALSE  NA    0.91.10
#>  4 rstudio-server-0.91.10-x86_64.rpm 2011-01-19 13:38:43 "\"826f8e2aca786536d938f81a6c415e1f\"" 5732549 STANDARD      FALSE  NA    0.91.10
#>  5 rstudio-server-0.91.12-amd64.deb  2011-01-20 18:19:05 "\"6372c1ebbd0736ada9243b44427ef6a8\"" 5513496 STANDARD      FALSE  NA    0.91.12
#>  6 rstudio-server-0.91.12-i386.deb   2011-01-20 18:19:06 "\"265ffb52d7d4191d6ee4e773c558fee9\"" 5446822 STANDARD      FALSE  NA    0.91.12
#>  7 rstudio-server-0.91.12-i686.rpm   2011-01-20 18:19:09 "\"21369000904044e6146c7f446d3843c2\"" 5611862 STANDARD      FALSE  NA    0.91.12
#>  8 rstudio-server-0.91.12-x86_64.rpm 2011-01-20 18:19:07 "\"78c31973947cd9059390c223335b5663\"" 5736415 STANDARD      FALSE  NA    0.91.12
#>  9 rstudio-server-0.91.13-amd64.deb  2011-01-20 20:06:07 "\"279787e2b126326197286b497224fd97\"" 5519202 STANDARD      FALSE  NA    0.91.13
#> 10 rstudio-server-0.91.13-i386.deb   2011-01-20 20:06:08 "\"a77fed41c254ac40e8c50f8a1155689b\"" 5451900 STANDARD      FALSE  NA    0.91.13
#> # ℹ 989 more rows